+++ syntheseus/reaction_prediction/inference/megan.py
@@ -6,10 +6,12 @@ Code: https://github.com/molecule-one/megan
 The original MEGAN code is released under the MIT license.
 """
 
+from __future__ import annotations
+
 import os
 import sys
 from pathlib import Path
-from typing import List, Union
+from typing import Any, Optional, Union
 
 from rdkit import Chem
 
@@ -20,6 +22,7 @@ from syntheseus.reaction_prediction.utils.inference import (
     get_unique_file_in_dir,
     process_raw_smiles_outputs,
 )
+from syntheseus.reaction_prediction.utils.misc import suppress_outputs
 
 
 class MEGANModel(BackwardReactionModel):
@@ -89,7 +92,7 @@ class MEGANModel(BackwardReactionModel):
     def get_parameters(self):
         return self.model.parameters()
 
-    def _mols_to_batch(self, inputs: List[Molecule]) -> List[Chem.Mol]:
+    def _mols_to_batch(self, inputs: list[Molecule]) -> list[Optional[Chem.Mol]]:
         from src.feat.utils import fix_explicit_hs
 
         # Inputs to the model are list of `rdkit` molecules.
@@ -102,37 +105,46 @@ class MEGANModel(BackwardReactionModel):
                 a.SetAtomMapNum(i + 1)
 
             try:
-                mol = fix_explicit_hs(mol)
+                input_batch.append(fix_explicit_hs(mol))
             except Exception:
-                # Sometimes `fix_explicit_hs` may fail with an `rdkit` error. In such cases we give
-                # up and use the molecule as-is.
-                # TODO(kmaziarz): Investigate these cases in more detail.
-                pass
-
-            input_batch.append(mol)
+                # MEGAN sometimes produces broken molecules containing C+ atoms which pass `rdkit`
+                # sanitization but fail in `fix_explicit_hs`. We block these here to avoid making
+                # predictions for them.
+                input_batch.append(None)
 
         return input_batch
 
-    def __call__(self, inputs: List[Molecule], num_results: int) -> List[BackwardPredictionList]:
+    def __call__(self, inputs: list[Molecule], num_results: int) -> list[BackwardPredictionList]:
         import torch
         from src.model.beam_search import beam_search
 
         # Get the inputs into the right form to call the underlying model.
         batch = self._mols_to_batch(inputs)
+        batch_valid = [mol for mol in batch if mol is not None]
+        batch_valid_idxs = [idx for idx, mol in enumerate(batch) if mol is not None]
 
-        with torch.no_grad():
-            beam_search_results = beam_search(
-                [self.model],
-                batch,
-                rdkit_cache=self.rdkit_cache,
-                max_steps=self.max_gen_steps,
-                beam_size=num_results,
-                batch_size=self.beam_batch_size,
-                base_action_masks=self.base_action_masks,
-                max_atoms=self.n_max_atoms,
-                reaction_types=None,
-                action_vocab=self.action_vocab,
-            )  # returns a list of `beam_size` results for each input molecule.
+        if batch_valid:
+            with torch.no_grad(), suppress_outputs():
+                beam_search_results = beam_search(
+                    [self.model],
+                    batch_valid,
+                    rdkit_cache=self.rdkit_cache,
+                    max_steps=self.max_gen_steps,
+                    beam_size=num_results,
+                    batch_size=self.beam_batch_size,
+                    base_action_masks=self.base_action_masks,
+                    max_atoms=self.n_max_atoms,
+                    reaction_types=None,
+                    action_vocab=self.action_vocab,
+                )  # returns a list of `beam_size` results for each input molecule
+        else:
+            beam_search_results = []
+
+        assert len(batch_valid_idxs) == len(beam_search_results)
+
+        all_outputs: list[list[dict[str, Any]]] = [[] for _ in batch]
+        for idx, raw_outputs in zip(batch_valid_idxs, beam_search_results):
+            all_outputs[idx] = raw_outputs
 
         return [
             process_raw_smiles_outputs(
@@ -140,5 +152,5 @@ class MEGANModel(BackwardReactionModel):
                 output_list=[prediction["final_smi_unmapped"] for prediction in raw_outputs],
                 kwargs_list=[{"probability": prediction["prob"]} for prediction in raw_outputs],
             )
-            for input, raw_outputs in zip(inputs, beam_search_results)
+            for input, raw_outputs in zip(inputs, all_outputs)
         ]